Split out MetNet preprocessing
