+++ satflow/models/layers/__init__.py
@@ -5,3 +5,4 @@ from .SpatioTemporalLSTMCell_memory_decoupling import SpatioTemporalLSTMCell
 from .RUnetLayers import Recurrent_block, RRCNN_block
 from .ConditionTime import ConditionTime
 from .Attention import SelfAttention, SelfAttention2d
+from .MetNetPreprocessor import MetNetPreprocessor
